# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Classroom Supplies Usage & Replenishment Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent supply activity
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
  /notifications/{id}/read:
    post:
      operationId: readNotification
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
...
  /reports/usage:
    get:
      operationId: getUsageReport
      tags: [reports]
      summary: Get usage report data
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [daily, weekly, monthly]
            default: daily
        - name: classroomId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Usage statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                    quantityUsed:
                      type: number
                    supplyName:
                      type: string
                    classroomName:
                      type: string
...
components:
  schemas:
...
    ActivityItem:
      type: object
      required: [id, supplyId, supplyName, quantityUsed, usedBy, usedAt, type]
      properties:
        id:
          type: integer
        supplyId:
          type: integer
        supplyName:
          type: string
        quantityUsed:
          type: number
        usedBy:
          type: string
        notes:
          type: ["string", "null"]
        usedAt:
          type: string
        type:
          type: string
          enum: [usage, restock]
...